requirements.txt
edgartools>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
requests>=2.31.0
